PROTON-214: fixed Messenger test race condition affecting the teardown and run methods.
In teardown(), the client sets running=false and sends a message (intended to
cause the server to promptly exit). If this completes before the server has
started (in run()), the server *never* starts, so the client send() times out.

This is particularly likely when tearing down testSendBogus() because the test
method completes very quickly.

I addressed this race condition using an Event to delay the client until the
server has decided to start.

This test is still racy because the client doesn't have a way of waiting until
the server has actually started so has to approximate it with a generous
timeout, but this is unavoidable given the current Messenger threading model.

git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/trunk@1459089 13f79535-47bb-0310-9956-ffa450edef68