PROTON-693: Python Url class to wrap C function pni_parse_url
It was pointed out that pni_parse_url is an internal function and the interface
is not suitable for public API.

Rewrote the URL parser as a proper swigable C API pn_url_*. This gets rid of the
need for previous swig insanity and is cleaner all round.

Internally still uses the pni_parse_url parser, we can clean that up later.

git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/branches/examples@1630836 13f79535-47bb-0310-9956-ffa450edef68