PROTON-752: Provide a non-blocking means to receive messages in Ruby.
To avoid changing the APIs or their intentions, a new class method is
added:

  Qpid::Proton::Receiver::receive_and_call

The method takes as arguments either an existing instance of Messenger
or else the parameters to create an instance, and also a code block to
be called with each message received.

The messenger is then put into passive mode and a new thread started. It
then monitors the messenger and, when a new message is received, passes
it to the block for processing.

To exit the messenger, the code would call Messenger#interrupt. The
thread then exits and the messenger stops processing.