NO-JIRA: proton.Url fill in  defaults by default.
Constructing a proton.Url will call self.defaults() to fill in any missing parts
of the url with defaults. This is almost always what you want and means that
users don't have to call url.defaults() manually.

The old behavior is still available by constructing the Url with defaults=False,
in case this is ever useful.  In that case we parse the URL string but leave any
missing fields as None. The str of the URL reflects the original, incomplete
string.