PROTON-772: New proton logging code
Based on list discussion:
- moved the log statement api into src/log_private
- simplified the public API to 2 functions: pn_log_enable and pn_log_logger

The minimal public API is required to let applications using proton control
where proton log output goes, which is the point of all this.

The present state of the logging achieves the following goals:
- allow applications to disable or redirect ALL proton library output, nothing is forced to stderr.
- leaves the previous behavior of transport tracers unaffected.
- introduces as little disturbance as possible (one new env. var PN_TRACE_LOG and 2 new API functions)

This is definitely not the final word on logging for proton, but a more
sophisticated, unified logging system requires more discussion in the proton
community.