NO-JIRA: Replace 0.0.0.0 with 127.0.0.1 connecting to a host in tests.
In some circumstances using 0.0.0.0 to connect fails. This has been observed when:
running the jython messenger tests, on fedroa 20, with an OpenVpn connection running.

In messenger.py, 0.0.0.0 has been replaced with 127.0.0.1 for connecting only, listening
to ~0.0.0.0 works fine and is unchanged.

The man page ip(7) suggests that 0.0.0.0 should only be used for bind:

       There  are  several special addresses: INADDR_LOOPBACK (127.0.0.1) always refers to the local
       host  via  the  loopback  device;  INADDR_ANY  (0.0.0.0)  means  any  address  for   binding;
       INADDR_BROADCAST  (255.255.255.255)  means  any  host  and  has  the  same  effect on bind as
       INADDR_ANY for historical reasons.

So we may need to replace 0.0.0.0 in other places, e.g. as default host value for the Url class.
This is not yet done pending further discussion.