PROTON-827: Start of work on Go binding - Go wrapper for pn_url_t type.
This commit creates a Go workspace with a package apache.org/proton containing
a Url type that uses cgo to wrap the pn_url_t functions. It is a placeholder for
future Go work. There is a README.md outlining the current thinking.

The go directory is not yet linked to the cmake system, it will be eventually.