PROTON-799: Added the pn_rbkey_t type to the Ruby APIs.
The pn_rbkey_t type provides a way, from Ruby, to attach pure Ruby
objects to Proton structs by means of an attachment.

The type holds a reference to an registry object, a method to invoke on
that object, and the value to pass to that method when invoked. This
method is then called when the rbkey is finalized, with the goal of
removing the referenced Ruby object so it can be safely garbage
collected.