NO-JIRA: Measure size of encoded data.
Introduce pn_data_encode2 which allows you to both encode data and, if you don't have enough
space, find out how much space you need. You can also use this to simply find out how
much space is required to  the data.

This supports two patterns of use:

1. Check how much space you need, allocate the space and encode.

2. Try to encode into a buffer you already have. If that doesn't work, re-allocate
   to the correct size and encode.

Eliminates the need for guessing and buffer doubling strategies, which are
inefficient given that we know exactly what is in the data.

The original pn_data_encode is unmodified for compatibility.