NO-JIRA: export.sh creates pax-based tar
Having switched to using git-archive, we end up creating a tar that uses
the PAX-based scheme for supporting long pathnames. This causes issues
for a number of extraction tools (ant, 7-zip, etc.) which only support
the GNU tar format. We should stick with GNU tar, as used for previous
qpid-proton releases.

Also update the script to revert to using a suffix of .tar.gz (rather
than .tgz) for the source distribution.