PROTON-881: Rough first pass at a proton-j -> proton-c interop. test.
This adds a Python test case which starts both a Python (proton-c)
reactor and also spawns a JVM running a (proton-j) reactor.  The
expected outcome is that the proton-j reactor is able to send a
message to the proton-c reactor.

Right now, there are a lot of rough edges on both the Pyton and
Java sides of this test case.