NO-JIRA: Update API doc for pn_delivery_settle, add public function pn_delivery_current
pn_delivery_settle() has an undocumented side effect: if the delivery is the
current delivery on its link, then the link is automatically advanaced.

Updated the API doc to explain this side effect, and also made public the helper
function pn_delivery_current() so callers can easily tell whether or not this
side effect will occur.