PROTON-881: Add code to release resources (e.g. sockets, selectors, etc.)
The proton-j codebase does not, generally, implement cleanup logic in
the same way as proton-c (explicit reference counting) and for some
resources (such as sockets, selectors, etc.) cannot always way for
garbage collection to finalize objects.  This commit adds a 'free'
method to a number of classes.  Calling this closes any Java resources
held by the class.  The reactor also has a 'free' method, which frees
both the reactor and any children that the reactor has.