PROTON-895: Rely on python to build downloaded tarball
Instead of using cmake to build the downloaded tarball, rely on python
for building and installing the library. This process is not only
cleaner but also plays nicer with the environment we're trying to give
support to.

In addition to the above, this plays nicer with the user and removes a
requirement for having the library installed. Note that the built
library will no longer overwrite system libraries but it'll be placed
along to the python bindings we're building.

(cherry picked from commit f2e97708a9c0da049bbc5089909fc31ab092edf8)