PROTON-908: Remove install-runtime dependency on swig
The patch allows python-qpid-proton for generating the `cproton_wrap.c`
file when creating the distribution package. Since this file is already
cross-platform, it should be safe to ship it as part of the
python-qpid-proton package.

This allows us for building the bindings even when swig is not
available. The setup.py process will try to generate a new
cproton_wrap.c everytime but it'll fallback to the one distributed with
the package if swig is not available.

Note that this will be done just in the platforms where bundling
proton-c is supported. The reason being that, in the other platforms, w
can't verify whether the installed qpid-proton version is the same that
was used to generated these files