PROTON-995: Fix pn_url_str to use URL hex escapes to generate a valid URL string.
pn_url_parse decodes URL encoded hexadecimal escape sequences in the username
and password, but pn_url_str was not re-encoding them resulting in an invalid
URL string if they contained the forbidden characters ':', '/' or '@'.

This fixes pn_url_str to hex-encode those characters in username and password fields.

PROTON-995 was reported against the python binding, the python Url class wraps
a pn_url_t so this also fixes the python problem.