PROTON-1075: Data races detected in go_test_electron.
Removed un-necessary code to drain the Engine.inject channel, the channel is not
buffered so there cannot be anything to drain. Inject also selects on the done
channel so all goroutines waiting to inject will unblock.

Made endpoint.closed() idempotent, fixing the above revealed that it can be
called twice depending on timing.