PROTON-1067: python messenger: cannot acknowledge messages, messenger forces auto-ack or pre-settled.
Change send-settle-mode defaults on outgoing links (but respect user settings if
there are any)

- sender: if outgoing_window, default to PN_SND_MIXED - let the application decide.
- reciever: if incoming_window, default to PN_SND_UNSETTLED - let the app settle.

Clearly if the app sets an incoming window, they want to settle messages
manually so requesting "UNSETTLED" by default makes no sense. It works with
existing messenger tests and examples because links are always established from
the sending end, but making a receiving (subscribing) link to a broker such as
qpidd shows the problem: messenger always requests SND_SETTLED, the broker
obliges, so it is impossible for the receiver to control settlement.