PROTON-1133: Policy for setting the virtual-host using the Reactor
The pn_connection_set_hostname() interface is used to set the
'hostname' field in the Open performative.  By definition this is the
'virtual host' and should not be used for the transport network address.

The network address for outgoing connections should be set
by using the reactor's pn_reactor_connection_to_host() factory, or the
pn_reactor_set_connection_host() when re-connecting to a different
host.

For inbound connections the peer address is provided by the acceptor
and cannot be modified.  In both cases, the
pn_reactor_get_connection_address() method can be used to obtain the
peer's network address.

By default if the application does not explicitly set the connection's
virtual host on an outbound connection Reactor will use the host
name/address from the network address.  Otherwise the virtual host is
sent as configured, with one exception: if the application sets the
virtual host to an empty string ("") then no virtual host will be sent
in the Open performative (e.g. disables the default behavior).

The Python Container now provides a virtual-host option for setting
the virtual host value via the connect() method.