PROTON-1236: add a link to the transport's reactor as soon as the it is bound
Previously the reactor would be put on the transport's attachment list
when a socket selector was created for the transport.  If the socket
setup fails then no selector is created and the reactor is never added
to the transport.  This would result in a crash as the code depends on
the reactor being present in the transport when transport events are
generated.

(cherry picked from commit 03b4ac60932110ffe0352195add7667dc11a2875)