Revert "PROTON-1258: Make transport events more consistent on client and server"
This reverts commit 4a6010110296846e3c77defc0558a45d8792d3b5.

This commit crashes the cpp_mt_example_test in a release build (but not in a
debug or reldbg build which is worth investigating separtely)

The connection engine relies on the events generated by correct use of
pn_transport_t. PROTON-1258 is a reactor bug: the reactor needs to be fixed to
close the transport correctly, before unbinding it, so that it can generate the
proper close events in the proper sequence. The handler should not be involved.