PROTON-1356: remove un-necessary codec_shim.h
PROTON-1356 reports problems with `go get`, with an error in types.go line 33,
pn_types_t not recognized. Not clear if this solves the problem, but types.go
was depending on a codec_shim.h file in the go package, which should have
included the pn_types_t definition. This shim is no longer needed and has been
removed.