PROTON-1419: c libuv proactor log all events
Use the pn_log() function to log listener and proactor events when PN_TRACE_LOG is set.

The current logging features are not ideal but this fix is consistent with the
status quo so if/when we upgrade the logging framework this will be included.

Add missing function definition for  pn_connection_driver_logf