NO-JIRA: Disable py33,py34 TOX environments in travis.yml
Make the tox.ini envlist settable from cmake option TOX_ENVLIST

Update .travis.yml to exclude py33,py34 from the envlist as both cause failures.

cmake prints a warning if TOX_ENVLIST is set to any non-default value to remind
us that this is a temporary measure.

When the problem is fixed, revert .travis.yml to re-enable the defaults.