PROTON-1460: Improved proactor selection in cmake.
Pick a proactor with `cmake -DPROACTOR=name` where name is 'epoll', 'libuv' or 'none'.
'none' disables the proactor even if a default is available.

PROACTOR="" (the default) picks the first proactor that has its build
requirements satisfied.

Presently 'epoll' is not included as a default. When it is ready, enable the
commented part of this line:

    if (PROACTOR STREQUAL "epoll") # OR (NOT PROACTOR AND NOT PROACTOR_OK))