PROTON-1495: c epoll proactor handling file descriptor shortage
Previously if the proactor ran out of file descriptors during pn_listener_accept the
listener closed, and could not accept further connections even when FDs became
free.

Now the affected listener(s) are put into "overflow" mode and dispabled until
the proactor closes an FD, at which point it re-enables overflowing listeners.

The error is reported as a PN_TRANSPORT_ERROR on the connection that the
application attempted to use for pn_listener_accept() Clients will experience a
delay and/or be disconnected while the listener is in overflow mode.

NOTE: we should also have a timeout to re-try overflow listeners even if the
proactor has not closed any file descriptors, in case other parts of the
application have. This is not yet done.