PROTON-1532: Allow insecure mechanism in SASL
This commit extends SASL with the additional allow_insecure_mechs allowing users to override the defailt (false) authentication and use the plain mechanism. It also extends the Connection class to expose user and password that is used by the plain mechanism.

Finally, this patch modifies the connect method which now avoids the use of the plain method which is not defined on SASL. The code now more resembles the Python version.

Signed-off-by: Gregor Berginc <gregor.berginc@gmail.com>