PROTON-1618: c++ tests use test_port.h for listen ports
POSIX: Use bind(0) with SO_REUSEADDR and hold the socket to acquire a port that
can safely be used for listen()

Windows: Use bind(0) to pick a port, but close the socket immediately. In theory
another process could steal the port between bind() and listen(), but in
practice this seems to be very unlikely.

The previous randomize-and-retry approach makes it hard to test the sequence of
events for listen(), since the random retry may cause multiple listen errors
even when it is finally successful.