PROTON-1369: [c++] remove deprecation attribute on void_function0
Was causing a deprecation warning on #include of work_queue because the the
deprecated function `work_queue::add(void_function0)` refers to the deprecated
class.

void_function0 is still deprecated in documentation and cannot be used except via
the deprecated work_queue::add function so effectively it is still well deprecated.