PROTON-1621: Reworked - C++11 example testing
Revert and rework the previous solution which pushed too much C++ detail to the
top level. Now each binding can export a small, set of selected extra flags for
examples that are built in a repo. These flags are optional, the examples must
be able to build without them when installed stand-alone. These flags are
exported as CACHE INTERNAL cmake variables for visibility, there is no
binding-specific cmake code outside the binding trees.