PROTON-1639 proton.c ipv6 test does not detect lack of ipv6 support
The test was checking for IPV6 by listening on "::" INADDR6_ANY.

On some IPV6-capable platforms with IPV6 disabled, this listen() succeeds even
though it not possible to connect to the listening port.

Modified the test to listen() on "::1" INADDR6_LOOPBACK. A host without ipv6
disabled should report this as an error as it will not have an IPV6 loopback
available.