PROTON-1639: epoll.c assert() on immediate connect() failure
The epoll proactor was asserting if a connect() call failed immediately.

This doesn't show up in most tests because the connect() is non-blocking, and
most failures are deferred to epoll and handled correctly

However trying to connect() to an IPv6 address with IPv6 disabled does fail
immediately and causes an assert() failure.