PROTON-1658: [ruby] simpler, more portable port allocation for tests
The SO_REUSEADDR trick that works on Linux does not work on OSX or Windows.
Use a simpler approach of listen(0), close socket, do real listen. This is
potentially race-prone but works well in practice

The correct solution is to allow the container to listen using a pre-existing socket,
this is coming with PROTON-1064.