PROTON-1537: [ruby] Split Delivery API into Tracker and Transfer.
Delivery is now used only for incoming messages, Tracker for outgoing messages,
and Transfer is the common base class for both.

Since most of the previous Deliveyr behavior is common, Delivery is still
backwards-compatible with some deprecated methods.

The API for release and modify is still compatible but has been re-worked so
that Delivery#release(opts) now handles release and modified states with
appropriate option settings. Tracker#modified provides to modified state
directly.

The Disposition class has is deprecated in favour of the new Delivery/Tracker methods.