PROTON-1537: [ruby] idle time-out conversion
Idiomatic ruby expresses intervals in seconds, which can be Float or Rational
for sub-second intervals.

Use seconds in the :idle_timeout connection options and Connection#idle_timeout.
Transport remains in milliseconds but is no longer a user-facing API.
Conversion is done using Rational to preserve millisecond accuracy, although
the user is free to use Float if rounding is not a concern.