PROTON-1762: [ruby] remove gem self-test
Test was causing mysterious problems on some platforms.
Isolating the installed gem and use the uninstalled qpid-proton library from the build
reqiures some hacking that appears not to be portable.

Opened https://issues.apache.org/jira/browse/PROTON-1765 to restore these tests
in the future if the problems can be solved.