PROTON-1766: [cpp] Remove thread-unsafe use of pn_record_t by proactor.
The proactor was using pn_record_t (via pn_connection_attachments) to navigate
from pn_connection_t to proactor data in arbitrary threads. This was unsafe, and
cannot be made safe without locks in the proton core because users can also use
pn_connection_attachments() outside the proactor's control.

Replaced use of attachments with a direct pointer field in pn_connection_t
which is used by the proactor to point to its internal data. Currently
access is locked by a mutex but this can/should be an atomic operation.