PROTON-1771: [c] locking around epoll_extended_t
Add locking for epoll_extended_t structure to fix race conditions revealed
by threaderciser.c

We pass pointers to this struct between threads via epoll, and epoll does not
guarantee that memory writes by one thread will be visible in the next thread
that gets the pointer. Since helgrind reports the races, it suggests that epoll
is not acting as a memory barrier in practice either.

These mutexes will not be contended so they are not expensive. We could use an
acquire/release memory barrier, but an un-contended mutex is basically exactly
that, and is portable.