Fix build of SASL related code on Windows
The code that uses Cyrus SASL library failed to build on Windows because
of "pthead.h" include. There is for sure a nicer way how to make
locking in a platform independent way. I have tried to keep the
pthread code for Unix and have added a corresponding code for Windows.