PROTON-2339: Proactor raw connection: Introduce DRAIN_BUFFERS event
This will supercede the use of READ/WRITTEN events to tell the
application to retrieve its buffers before DISCONNECT.

This more specific event should be clearer in purpose.

For backwards compatibility with existing code, if the application
doesn't drain the buffers in response to the drain event, we will
send the read/written events (as applicable) as well. This behaviour
should be removed when the existing major application (dispatch) is
fixed to use this new event.