PROTON-2354: Fix to the new destructors that works with C++03
We do still support C++03 and the previous fix had the wrong exception
specification for pre C++11 destructors:

The base exception (std::runtime_error) is specified to not throw any
exeptions so we need to specify the new destructors the same way.
C++11 defaults to non throwing destructors so omitting the exception
specification works there. However earlier C++ doesn't default like this
and so we need to do this explicitly for C++03. The explicit specification
works fine for C++11 as it doesn't actually change anything there.