NO-JIRA: [Python] Fix exception in IOHandler.on_selectable_expired/update
This occasionally happens after my laptop wakes up from overnight sleep.

IOHandler.on_selectable_expired() is invoked with `selectable` that has
`_terminated=True` and `_transport=None` so when `IOHandler.update()`
is called it crashes when trying to handle the exception
(since transport is None).

Fix this by checking if transport attribute is set and that selectable is not
terminated before invoking `IOHandler.update()`.

Signed-off-by: Ievgen Popovych <ievgenp@seetrue.ai>