PROTON-2542: Simplify pn_class_t handling
Allow every vtable entry to be NULL and have a good default behaviour.
This allows a lot of devirtualization as the majority of the entries are
now NULL and are handled without an indirect jump. It also reduces the
number of symbols needed to be exported to be in the pn_class_t tables.