PROTON-2542: Remove last vestiges of now unnecessary "reify" operation
We have now removed every previous case that relied on the object system
looking up the actual class of an object and this is now entirely
delegated to the objct classes themselves. This means we can remove the
generic code that looks up the object class in all of the general paths.