PROTON-2542: Make PN_OBJECT generic object management
Rename the previous PN_OBJECT to PN_DEFAULT as it contains the default
handling for all object based classes. PN_OBJECT is now a memory
management strategy used for containers which incorporates the behaviour
that was previously implemented by reify.

This means that PN_OBJECT can be used as the container class type for
generic objects.