PROTON-2557: Introduce pn_tostring in place of pn_inspect
Introduce and use a new API pn_tostring which produces a string
representation of a Proton object in a char* string. It allocates this
string with malloc and the application must free it.

This replaces pn_inspect which requires the application to use the
pn_string_t type purely to receive the string output. But we really
don't want or need to expose this type outside the library itself.