PROTON-2645: Allow pip to install python binding on Windows
The issue seems to be that the python setup.py uses absolute paths to
refer to the C source that is being compiled and these long paths get
propagated to the compiler to refer to the files. In some configurations
the absolute path is too long for the system limits on path length and
the compile fails.

This fix attempts to construct a much shorter relative path to the files
to compile and so not fall foul of the path length restrictions.