PROTON-2756: Allow python testing to use the python detected by CMake
This is comtrolled by a new CMake option called ENABLE_PYTHON_TEST_VENV

If the build system is not connected to the internet then building a
virtual environment to test the python binding can be a problem so allow
the build to use the installation of python that was detected by CMake
to test the python binding.

Testing using a completely clean environment made for the purpose will
be more reliable for internet connected systems so using the venv is the
default.

Since the switch to using cffi instead of swig this is required for
testing to work and for building wheels. So we make sure that we have
cffi installed before we allow python testing or wheel building.