PROTON-2816/PROTON-2817: Support building python extension with unbundled proton
There are 3 cases:
* Bundled proton library - this should be the default and is what you
  should use with pip etc.
* System proton library - this is what you should use in a distro
  package.
* Use the in tree library built earlier in the build process - this is
  what you usually want for developer debugging, and should be the
  version actually tested by the in tree tests.

This change also add control of the ability to package and test in
isolated virtual python environments. Using these is the default, but
under some circumstances you do not want the build process to construct
a virtual environment by pulling packages from the internet. So we now
have an option to turn this behaviour off and rely on the hosts python
environment.