PROTON-2594: c/examples/broker: Fix error message
The error message unconditionally mentions tserver-private-key.pem,
which is not applicable to Windows, where tserver-full.p12 is used
instead. Fix this by using a new PRIVATEKEY macro that expands to the
correct file name depending on whether we compile for Windows or not.

This prepares us for adding a third PKCS#11 #ifdef branch in the
follow-up commit, which is also why we are moving SSL_FILE specific
into the existing #ifdef branches.