PROTON-2594: c/examples/broker: add PKCS11_URI support
The broker hardcodes the name of the files used at compile time,
depending on whether it's built for Windows or not.

Add a third option by checking whether PKCS11_URI is defined: If it is,
it will use it to reference a "tserver" certificate and private key that
should be used instead of the keys in the ssl-certs/ directory.

This allows an easy manual test of the PKCS#11 functionality.