NO-JIRA: c/cpp: fix examples installation
The c/cpp examples so far installed the source files into ${PROTON_SHARE}.
This is not what most users would expect, especially if we are
cross-compiling and the target lacks a toolchain.

Fix this by installing the binary artifacts instead. For them to work
out-of-the-box, we additionally copy the ssl-certs to the expected
location as well.