PROTON-2594: [C++] connect_config_test: factor out some definitions into header
The test_handler implementation in connect_config_test can be useful for
further tests as well such as the incoming pkcs11_test.

We don't want to add the PKCS#11 test here, because the test is not
applicable to all platforms and it would overly complicate the existing
tests, so let's share code via a common header.