PROTON-2594: ci: setup environment for PKCS#11 test
Now that all pieces are in place, let's adapt out Github Action
workflow, so softhsm2, opensc (for pkcs11-tool) and pkcs11-provider
(for OpenSSL integration) are available and amend the ctest step
to source scripts/prep-pkcs11_test.sh, so the relevant environment
variables are exported.

Note that pkcs11-provider is shipped in ubuntu-24.04, but it doesn't
work, so we build it from source instead. We still need to use
ubuntu-24.04, because otherwise libcrypto would be outdated:
>= 3.0.7 is required, but ubuntu-latest is 22.04.5, which ships
libcrypto 3.0.2.