PROTON-2898: [C] Introduce pn_proactor_import_socket API
Implement functionality to import a socket that is already connected to
the proactor.

This can be for either the server or client end of the
connection. So it can be used for both socket activation at the server
end and protocols running on the connection prior to AMQP at the
client end.

The new pn_proactor_import_socket API is defined in a new header
proactor_ext.h which will contain any other similarly OS dependent
proactor APIs.