Corrected handling of outcomes
However there is still a problem for handling errors:
The spec requires a rejected disposition or detaching the link to
indicate an error with the declare or commit. Rollback cannot error as
the outcome would be the same!

In this current code if the outcome is not valid it will throw an
unhandled exception - this is not sufficiently robust behaviour!

Added in handling for errors so that the application can inspect the
transaction error in the rejected disposition.