PROTON-2873: [Python] Allow Delivery.update to take Disposition
To allow more convenient local delivery disposition updates allow update
to take a LocalDisposition object and use that for the update as well as
a disposition type.

Also allow TransactionalDisposition to take a Disposition outcome as
well as a disposisiton type for convenience of creating a more complex
transactional disposition update. At this point the underlying C API
only allows the outcome type to be recorded though.